chore: refresh generated artifacts (repo-index, graph, api-map) [T000607]
