+++ brett/src/client/mannequin.ts
@@ -594,59 +594,3 @@ export function clearModerationVisuals(figures: any[]): void {
   updateModerationVisuals(figures, { spotlight: null, dim: null, freeze: false });
 }
 
-// ── Filter Visuals (T000607) ───────────────────────────────────────────────
-
-const FILTER_DIM_OPACITY = 0.15;
-
-/**
- * Per-frame filter visual updater. Dims non-matching figures to FILTER_DIM_OPACITY
- * and restores matching (or all, when query is empty) figures to opacity 1.
- *
- * Called from the board-boot tick loop AFTER updateModerationVisuals so that
- * moderation takes precedence: if moderation already dimmed a figure, this does
- * not raise its opacity.
- *
- * The function intentionally does NOT use the _moderationCache because filter
- * state is local and does not need DB-roundtrip stability.
- */
-export function updateFilterVisuals(figures: any[], query: string): void {
-  const hasFilter = query.length > 0;
-
-  for (const fig of figures) {
-    const matches = !hasFilter || (fig.label ?? '').toLowerCase().includes(query.toLowerCase());
-
-    fig.root.traverse((o: any) => {
-      if (!o.isMesh || !o.material) return;
-      if (o === fig.ring || o === fig.possessionRing) return;
-      if (o.userData?.isContact) return;
-
-      if (!hasFilter) {
-        // Restore — only if we previously dimmed via filter (avoid fighting moderation)
-        if (o.material._filterDimmed) {
-          o.material.opacity = o.material._filterOriginalOpacity ?? 1.0;
-          o.material.transparent = o.material._filterOriginalTransparent ?? false;
-          o.material._filterDimmed = false;
-          o.material.needsUpdate = true;
-        }
-      } else if (!matches) {
-        // Dim non-matching figure
-        if (!o.material._filterDimmed) {
-          o.material._filterOriginalOpacity = o.material.opacity;
-          o.material._filterOriginalTransparent = o.material.transparent;
-          o.material._filterDimmed = true;
-        }
-        o.material.opacity = FILTER_DIM_OPACITY;
-        o.material.transparent = true;
-        o.material.needsUpdate = true;
-      } else {
-        // Matching figure — restore if previously dimmed by filter
-        if (o.material._filterDimmed) {
-          o.material.opacity = o.material._filterOriginalOpacity ?? 1.0;
-          o.material.transparent = o.material._filterOriginalTransparent ?? false;
-          o.material._filterDimmed = false;
-          o.material.needsUpdate = true;
-        }
-      }
-    });
-  }
-}
