+++ docs/code-quality/repo-index.json
@@ -2333,7 +2333,7 @@
       "id": "brett",
       "name": "Brett 3D board (CommonJS)",
       "owner_agent": "bachelorprojekt-website",
-      "file_count": 296,
+      "file_count": 298,
       "files": [
         "brett/.dockerignore",
         "brett/.gitignore",
@@ -2514,6 +2514,7 @@
         "brett/src/client/ui/skin.ts",
         "brett/src/client/ui/theme.ts",
         "brett/src/client/ui/timeline.ts",
+        "brett/src/client/ui/topbar-filter.ts",
         "brett/src/client/ui/topbar-invite.ts",
         "brett/src/client/ui/topbar-participants.ts",
         "brett/src/client/ui/undo-redo-ui.ts",
@@ -2622,6 +2623,7 @@
         "brett/test/theme.test.ts",
         "brett/test/tick-loop-priority.test.ts",
         "brett/test/timeline.test.ts",
+        "brett/test/topbar-filter.test.ts",
         "brett/test/topbar-invite.test.ts",
         "brett/test/topbar-participants.test.ts",
         "brett/test/undo-redo-hud.test.ts",
