+++ k3d/docs-content-built/architecture/index.html
@@ -178,7 +178,7 @@
 <body>
   <div class="header">
     <h1>⬡ Architektur — Living Docs</h1>
-    <span class="meta">Generiert: 2026-06-11T00:54:21.435Z</span>
+    <span class="meta">Generiert: 2026-06-11T01:08:21.527Z</span>
   </div>
 
   <div class="stats">
